+++ checksum-Elixir.Parquex.Native.exs
@@ -1,9 +1,0 @@
-%{
-  "libparquex_nif-v0.3.0-nif-2.16-aarch64-apple-darwin.so.tar.gz" => "sha256:b0913698823361750838ef8d5c9d02dd525beb78f6902cc18f79ef7080e78d66",
-  "libparquex_nif-v0.3.0-nif-2.16-aarch64-unknown-linux-gnu.so.tar.gz" => "sha256:666af019cbd6826f00d3d5f2eaf76f13d7d004f86890f6e0544754e6aca1b8f4",
-  "libparquex_nif-v0.3.0-nif-2.16-aarch64-unknown-linux-musl.so.tar.gz" => "sha256:0641cbddff5b221f8279dd3bfc427f395a002a411ed9005a70bdcf1c3060431c",
-  "libparquex_nif-v0.3.0-nif-2.16-x86_64-apple-darwin.so.tar.gz" => "sha256:52b9120bad04721a4c33bee08da2ca151a5849e7fa00d6f1de3e93c2382a738b",
-  "libparquex_nif-v0.3.0-nif-2.16-x86_64-unknown-linux-gnu.so.tar.gz" => "sha256:a5e81352c086b9a45e771eca7cdc4e324d6f5732b4469f87c9d647010cd4ede0",
-  "libparquex_nif-v0.3.0-nif-2.16-x86_64-unknown-linux-musl.so.tar.gz" => "sha256:cb0afef8764cc720b9f5ca00aedf1dc89767befc1f14280ac6ea332171b5ea5b",
-  "parquex_nif-v0.3.0-nif-2.16-x86_64-pc-windows-msvc.dll.tar.gz" => "sha256:47c60f36dbb79f47d5704aa39b593fec6226dc62a8f5de5c78989f13b6aaadd4",
-}